model-management-service/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
redis==5.0.1
aioredis==2.0.1
sqlalchemy==2.0.23
asyncpg==0.29.0
psycopg2-binary==2.9.9
httpx==0.25.2
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
sentence-transformers>=2.3.0
huggingface-hub>=0.20.0
numpy==1.24.3
tiktoken==0.5.2
websockets==12.0
python-socketio==5.10.0
aiofiles==23.2.1
structlog==23.2.0
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0

